backend/requirements.txt
quart
quart-cors
aiohttp
python-dotenv
hypercorn
